+++ base/base.h
@@ -86,7 +86,7 @@
 #include "cdrom/cdrom.h"
 #endif
 
-#define STD_BUFFER1_SIZE  ((long)61 * 1024)      /* weniger geht nicht! (Bild + CMAP)*/
+#define STD_BUFFER1_SIZE  ((int32_t)61 * 1024)      /* weniger geht nicht! (Bild + CMAP)*/
 
 extern char prgname[255];
 extern void *StdBuffer1;
